// src/enums/DocumentTypeRequest.ts
export enum DocumentTypeRequest {
  DEBTOR = 'DEBTOR',
  DOCUMENTCREDITCHECKING = 'DOCUMENT_CREDIT_CHECKING',
  DOCUMENTREVIEWERDELST = 'DOCUMENT_REVIEWER_DELST',
  DOCUMENTREVIEWERDEPI = 'DOCUMENT_REVIEWER_DEPI',
  DOCUMENTREVIEWERDH = 'DOCUMENT_REVIEWER_DH',
  DOCUMENTREVIEWERDK = 'DOCUMENT_REVIEWER_DK',
  DOCUMENTSITEVISIT = 'DOCUMENT_SITE_VISIT',
  DOCUMENTTECHNICALREVIEW = 'DOCUMENT_TECHNICAL_REVIEW',
  DOCUMENTTECHNICALREVIEWDELST = 'DOCUMENT_TECHNICAL_REVIEW_DELST',
  DOCUMENTUPLOADPKPT = 'DOCUMENT_UPLOAD_PKPT',
  LPA = 'LPA',
  LPAREVIEW = 'LPA_REVIEW',
  MANAGEMENT = 'MANAGEMENT',
  MEMOSUPPLEMENT = 'MEMO_SUPPLEMENT',
  OTHERRELATED = 'OTHER_RELATED',
  PKPTEFFECTIVETERMS = 'PKPT_EFFECTIVE_TERMS',
  PKPTSIGNINGTERMS = 'PKPT_SIGNING_TERMS',
  RISALAHRAPAT = 'RISALAH_RAPAT',
  RISALAHRAPATMERGED = 'RISALAH_RAPAT_MERGED',
  SHAREHOLDER = 'SHAREHOLDER',
  SPFP = 'SPFP',
  MIP = 'MIP',
  MIPANALYST = 'MIP_ANALYST',
  MUP = 'MUP',
  MUPANALYST = 'MUP_ANALYST',
  MIPREVIEW = 'MIP_REVIEW',
  DOCUMENTDEBTOR = 'DOCUMENT_DEBTOR',
  BENEFICIALOWNER = 'BENEFICIAL_OWNER',
  CUSTOMERDUEDILIGENCE = 'CUSTOMER_DUE_DILIGENCE',
  BAR = 'BAR',
  ADDITIONALDOCUMENT = 'ADDITIONAL_DOCUMENT',
  HIGH_RISK_CONCLUSION = 'HIGH_RISK_CONCLUSION',
}