// src/enums/Module.ts
export enum TypeModule {
  BAR = 'BAR',
  FAST_TRACK = 'FAST_TRACK',
  APU_PPT = 'APU_PPT',
  ASPECT_LEGAL = 'ASPECT_LEGAL',
  CREDIT_CHECKING = 'CREDIT_CHECKING',
  CREDIT_CHECKING_DOCUMENT = 'CREDIT_CHECKING_DOCUMENT',
  CREDIT_CHECKING_REQUEST = 'CREDIT_CHECKING_REQUEST',
  CREDIT_CHECKING_RESULT = 'CREDIT_CHECKING_RESULT',
  DPOP = 'DPOP',
  ENGAGEMENT_AGREEMENT = 'ENGAGEMENT_AGREEMENT',
  HIGH_RISK = 'HIGH_RISK',
  LPA = 'LPA',
  MAINTENANCE_DEBTOR = 'MAINTENANCE_DEBTOR',
  MASTER_DATA = 'MASTER_DATA',
  MIP = 'MIP',
  MIP_REVIEW = 'MIP_REVIEW',
  MUP = 'MUP',
  PIPELINE = 'PIPELINE',
  RISALAH_RAPAT = 'RISALAH_RAPAT',
  SITE_VISIT = 'SITE_VISIT',
  SPFP = 'SPFP',
  TECHNICAL_REVIEW = 'TECHNICAL_REVIEW',
  MAINTENANCE_GROUP = 'MAINTENANCE_GROUP',
  USER_MANAGEMENT = 'UM',
  BMPP = 'BMPP',
  LPS = 'LPS',
  MAINTENANCE_DATA = 'MAINTENANCE_DATA',
  BMPP_MONITORING = 'BMPP_MONITORING',
  MAINTENANCE_CAPITAL = 'MAINTENANCE_CAPITAL',
  MAINTENANCE_REMINDER = 'RMD',
  MAINTENANCE_NOTIFICATION = 'NTF',
  ACCESS_MENU = 'AM',
  VA_CREATION = 'VIRTUAL_ACCOUNT',
  PARAMETER_SLA = 'PARAMETER_SLA',
  PARAMETER_COT_EOD = 'PARAMETER_COT_EOD',
  PARAMETER_RATE = 'PARAMETER_RATE',
  PARAMETER_BENEFICIAL_OWNER = 'PARAMETER_BENEFICIAL_OWNER',
  PARAMETER_BUSINESS_CALL = 'PARAMETER_BUSINESS_CALL',
  PARAMETER_CUSTOMER_DUE_DILIGENCE = 'PARAMETER_CUSTOMER_DUE_DILIGENCE',
  PARAMETER_VA = 'PARAMETER_VA',
  PARAMETER_LOV = 'PARAMETER_LOV',
  PARAMETER_APU_PPT = 'PARAMETER_DOCUMENT_VERIFICATION',
  PARAMETER_SKEMA_SYARIAH = 'PARAMETER_SKEMA_SYARIAH',
  REASSIGNMENT_SKU = 'REASSIGNMENT_SKU',
  PARAMETER_SYARIAH = 'PARAMETER_SYARIAH',
  ANNUAL_REVIEW = 'ANNUAL_REVIEW',
  MONITORING = 'MONITORING',
}

export enum TypeProcess {
  BAR = 'BAR',
  FAST_TRACK = 'FAST_TRACK',
  MAINTENANCE_CUSTOMER = 'MAINTENANCE_CUSTOMER',
  APU_PPT = 'APU_PPT',
  APU_PPT_DPOP = 'APU_PPT_DPOP',
  ASPECT_LEGAL = 'ASPECT_LEGAL',
  ASSIGNMENT = 'ASSIGNMENT',
  CREDIT_CHECKING = 'CREDIT_CHECKING',
  CREDIT_CHECKING_DPOP = 'CREDIT_CHECKING_DPOP',
  CREDIT_CHECKING_MONITORING = 'CREDIT_CHECKING_RESULT|CREDIT_CHECKING_SUMMARY',
  CREDIT_CHECKING_REQUEST = 'CREDIT_CHECKING_REQUEST',
  CREDIT_CHECKING_RESULT = 'CREDIT_CHECKING_RESULT',
  CREDIT_CHECKING_SUMMARY = 'CREDIT_CHECKING_SUMMARY',
  DOCUMENT_VERIFICATION = 'DOCUMENT_VERIFICATION',
  ENGAGEMENT_AGREEMENT = 'ENGAGEMENT_AGREEMENT',
  HIGH_RISK = 'HIGH_RISK',
  HIGH_RISK_DK = 'HIGH_RISK_DK',
  LEGAL_SIGNING = 'LEGAL_SIGNING',
  LPA = 'LPA',
  LPA_REVIEW = 'LPA_REVIEW',
  MAINTENANCE_DEBTOR = 'MAINTENANCE_DEBTOR',
  MIP = 'MIP',
  MIP_ANALYST = 'MIP_ANALYST',
  MIP_REVIEW = 'MIP_REVIEW',
  MIP_REVIEW_REVISION = 'MIP_REVIEW_REVISION',
  MUP = 'MUP',
  MUP_ANALYST = 'MUP_ANALYST',
  PIPELINE = 'PIPELINE',
  PROCESSING_TYPE_PK = 'PROCESSING_TYPE_PK',
  REQUEST = 'REQUEST',
  REVIEWER_DELST = 'REVIEWER_DELST',
  REVIEWER_DEPI = 'REVIEWER_DEPI',
  REVIEWER_DH = 'REVIEWER_DH',
  REVIEWER_DK = 'REVIEWER_DK',
  REVIEWER_DLST = 'REVIEWER_DLST',
  RISALAH_RAPAT = 'RISALAH_RAPAT',
  SITE_VISIT = 'SITE_VISIT',
  SPDP = 'SPFP_DPOP',
  SPFP = 'SPFP',
  SPFP_FINAL = 'SPFP_FINAL',
  SPFP_REVISION = 'SPFP_REVISION',
  SPFP_CREATION_FINAL = 'SPFP_CREATION_FINAL',
  SUMMARY = 'SUMMARY',
  TECHNICAL_REVIEW = 'TECHNICAL_REVIEW',
  TECHNICAL_REVIEW_DELST = 'TECHNICAL_REVIEW_DELST',
  UPLOAD_RESULT = 'UPLOAD_RESULT',
  MAINTENANCE_GROUP = 'MAINTENANCE_GROUP',
  MAINTENANCE_REMINDER = 'RMD',
  MAINTENANCE_REMINDER_DETAIL = 'MAINTENANCE_REMINDER_DETAIL',
  MAINTENANCE_NOTIFICATION = 'NTF',
  USER_MANAGEMENT = 'UM',
  BMPP = 'BMPP',
  LPS_BAST = 'LPS_BAST',
  LPS_BAST_DPOP = 'LPS_BAST_DPOP',
  LPS_CORE = 'LPS_CORE',
  MAINTENANCE_MODAL = 'MAINTENANCE_MODAL',
  MAINTENANCE_CAPITAL = 'MAINTENANCE_CAPITAL',
  BMPP_MONITORING = 'BMPP_MONITORING',
  ACCESS_MENU = 'AM',
  MAINTENANCE_PROYEK = 'MAINTENANCE_PROJECT',
  VA_CREATION = 'VA_CREATION',
  PARAMETER_SLA = 'PARAMETER_SLA',
  PARAMETER_COT_EOD = 'PARAMETER_COT_EOD',
  PARAMETER_RATE = 'PARAMETER_RATE',
  PARAMETER_BENEFICIAL_OWNER = 'PARAMETER_BENEFICIAL_OWNER',
  PARAMETER_BUSINESS_CALL = 'PARAMETER_BUSINESS_CALL',
  PARAMETER_CUSTOMER_DUE_DILIGENCE = 'PARAMETER_CUSTOMER_DUE_DILIGENCE',
  PARAMETER_VA = 'PARAMETER_VA',
  PARAMETER_LOV = 'PARAMETER_LOV',
  PARAMETER_APU_PPT = 'PARAMETER_DOCUMENT_VERIFICATION',
  PARAMETER_SKEMA_SYARIAH = 'PARAMETER_SKEMA_SYARIAH',
  REASSIGNMENT_SKU = 'REASSIGNMENT_SKU',
  PARAMETER_SYARIAH = 'PARAMETER_SYARIAH',
  ANNUAL_REVIEW_REQUEST = 'ANNUAL_REVIEW_REQUEST',
  ANNUAL_REVIEW_ANALYST = 'ANNUAL_REVIEW_ANALYST',
  ANNUAL_REVIEW_ASSIGNMENT = 'ANNUAL_REVIEW_ASSIGNMENT',
  ANNUAL_REVIEW_VERIFICATION = 'ANNUAL_REVIEW_VERIFICATION',
  ANNUAL_REVIEW = 'ANNUAL_REVIEW',
  ANNUAL_REVIEW_DEPI = 'ANNUAL_REVIEW_DEPI',
  MONITORING = 'MONITORING',

}